requirements.txt
Flask==3.0.0
boto3==1.34.14
python-dotenv==1.0.0
psycopg2-binary==2.9.9
uuid==1.30
Pillow==10.1.0
requests==2.31.0